- 테이블 관련 코드 리팩토링(중복코드 정리)
