+++ com.hangum.tadpole.rdb.core/src/com/hangum/tadpole/rdb/core/editors/main/composite/direct/QueryResultLabelProvider.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     hangum - initial API and implementation
  ******************************************************************************/
-package com.hangum.tadpole.engine.sql.util.tables;
+package com.hangum.tadpole.rdb.core.editors.main.composite.direct;
 
 import java.util.HashMap;
 
@@ -22,11 +22,11 @@ import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Image;
 
 import com.hangum.tadpole.commons.util.NumberFormatUtils;
-import com.hangum.tadpole.engine.query.dao.system.UserInfoDataDAO;
 import com.hangum.tadpole.engine.sql.util.RDBTypeToJavaTypeUtils;
 import com.hangum.tadpole.engine.sql.util.resultset.ResultSetUtilDTO;
-import com.hangum.tadpole.preference.define.PreferenceDefine;
-import com.hangum.tadpole.session.manager.SessionManager;
+import com.hangum.tadpole.engine.utils.EditorDefine;
+import com.hangum.tadpole.engine.utils.EditorDefine.QUERY_MODE;
+import com.hangum.tadpole.preference.get.GetPreferenceGeneral;
 import com.swtdesigner.SWTResourceManager;
 
 /**
@@ -35,31 +35,42 @@ import com.swtdesigner.SWTResourceManager;
  * @author hangum
  *
  */
-public class SQLResultLabelProvider extends LabelProvider implements ITableLabelProvider, ITableColorProvider {
+public class QueryResultLabelProvider extends LabelProvider implements ITableLabelProvider, ITableColorProvider {
 	/**
 	 * Logger for this class
 	 */
-	private static final Logger logger = Logger.getLogger(SQLResultLabelProvider.class);
-	private boolean isPretty = false;
-	private String strNullValue = "";
+	private static final Logger logger = Logger.getLogger(QueryResultLabelProvider.class);
+	
+	private EditorDefine.QUERY_MODE queryMode = QUERY_MODE.QUERY;
 	private ResultSetUtilDTO rsDAO;
 	
-	public SQLResultLabelProvider() {
+	public QueryResultLabelProvider() {
 	}
 
-	public SQLResultLabelProvider(final boolean isPretty, final String strNullValue) {
-		this.isPretty = isPretty;
-		this.strNullValue = strNullValue;
-	}
-	
-	public SQLResultLabelProvider(final boolean isPretty, final ResultSetUtilDTO rsDAO, final String strNullValue) {
-		this.isPretty = isPretty;
-		this.strNullValue = strNullValue;
+	public QueryResultLabelProvider(EditorDefine.QUERY_MODE queryMode, final ResultSetUtilDTO rsDAO) {
+		this.queryMode = queryMode;
 		this.rsDAO = rsDAO;
 	}
 
 	@Override
 	public Color getForeground(Object element, int columnIndex) {
+		HashMap<Integer, Object> rsResult = (HashMap<Integer, Object>)element;
+		Object obj = rsResult.get(columnIndex);
+		if(obj == null) {
+			return SWTResourceManager.getColor(152, 118, 137);
+		} else {
+			String objValue = obj.toString();
+			
+			if(queryMode == QUERY_MODE.QUERY) {
+				if(GetPreferenceGeneral.getRDBShowInTheColumn() != -1 
+						&& objValue.length() > GetPreferenceGeneral.getRDBShowInTheColumn()
+						&& !RDBTypeToJavaTypeUtils.isNumberType(rsDAO.getColumnType().get(columnIndex))
+				) {
+					return SWTResourceManager.getColor(152, 118, 137);
+				}
+			}
+		}
+		
 		return null;
 	}
 
@@ -69,36 +80,29 @@ public class SQLResultLabelProvider extends LabelProvider implements ITableLabel
 		
 		return null;
 	}
-
+	
 	public Image getColumnImage(Object element, int columnIndex) {
 		return null;
 	}
-	
+
 	@SuppressWarnings("unchecked")
 	public String getColumnText(Object element, int columnIndex) {
 		HashMap<Integer, Object> rsResult = (HashMap<Integer, Object>)element;
 		
 		Object obj = rsResult.get(columnIndex);
-		if(rsDAO != null) {
-			if(isPretty & RDBTypeToJavaTypeUtils.isNumberType(rsDAO.getColumnType().get(columnIndex))) return NumberFormatUtils.addComma(obj);
-		}
-		String showValue = "";
-		try {
-			int intShowWidth = Integer.parseInt(getRDBShowInTheColumn());
-			if(intShowWidth != -1) showValue = StringUtils.abbreviate(obj.toString(), 0, intShowWidth);
-			else showValue = obj.toString();
-		} catch(Exception e) {}
 		
-		return obj == null ? strNullValue : showValue;
-	}
-	
-	/**
-	 * RDB Character shown in the column
-	 * @return
-	 */
-	private static String getRDBShowInTheColumn() {
-		UserInfoDataDAO userInfo = SessionManager.getUserInfo(PreferenceDefine.RDB_CHARACTER_SHOW_IN_THE_COLUMN, PreferenceDefine.RDB_CHARACTER_SHOW_IN_THE_COLUMN_VALUE);
-		return userInfo.getValue0();
+		if(obj == null) {
+			return GetPreferenceGeneral.getResultNull();
+		} else if(GetPreferenceGeneral.getISRDBNumberIsComma() && RDBTypeToJavaTypeUtils.isNumberType(rsDAO.getColumnType().get(columnIndex))) {
+			return NumberFormatUtils.addComma(obj);
+		} else {
+			if(GetPreferenceGeneral.getRDBShowInTheColumn() == -1) {
+				return obj.toString();
+			} 
+			return queryMode == QUERY_MODE.QUERY ? 
+					StringUtils.abbreviate(obj.toString(), 0, GetPreferenceGeneral.getRDBShowInTheColumn()) :
+					obj.toString();
+		}
 	}
 	
 }
