+++ com.hangum.tadpole.rdb.core/src/com/hangum/tadpole/rdb/core/editors/main/composite/plandetail/mysql/MySQLExtensionViewDialog.java
@@ -35,7 +35,7 @@ import com.hangum.tadpole.engine.sql.util.resultset.TadpoleResultSet;
 import com.hangum.tadpole.engine.sql.util.tables.SQLResultSorter;
 import com.hangum.tadpole.engine.sql.util.tables.TableUtil;
 import com.hangum.tadpole.engine.utils.RequestQuery;
-import com.hangum.tadpole.rdb.core.editors.main.composite.direct.SQLResultLabelProvider;
+import com.hangum.tadpole.rdb.core.editors.main.composite.direct.QueryResultLabelProvider;
 import com.swtdesigner.SWTResourceManager;
 
 /**
@@ -142,9 +142,9 @@ public class MySQLExtensionViewDialog extends Dialog {
 		final TadpoleResultSet trs = showProfiles.getDataList();
 		final SQLResultSorter sqlSorter = new SQLResultSorter(-999);
 		
-		SQLResultLabelProvider.createTableColumn(reqQuery, tvShowProfiller, showProfiles, sqlSorter, false);
+		TableUtil.createTableColumn(tvShowProfiller, showProfiles, sqlSorter);
 		
-		tvShowProfiller.setLabelProvider(new SQLResultLabelProvider(reqQuery.getMode(), showProfiles));
+		tvShowProfiller.setLabelProvider(new QueryResultLabelProvider(reqQuery.getMode(), showProfiles));
 		tvShowProfiller.setContentProvider(new ArrayContentProvider());
 		
 		// 쿼리를 설정한 사용자가 설정 한 만큼 보여준다.
@@ -165,9 +165,9 @@ public class MySQLExtensionViewDialog extends Dialog {
 		final TadpoleResultSet trs = showStatus.getDataList();
 		final SQLResultSorter sqlSorter = new SQLResultSorter(-999);
 		
-		SQLResultLabelProvider.createTableColumn(reqQuery, tvShowStatus, showStatus, sqlSorter, false);
+		TableUtil.createTableColumn(tvShowStatus, showStatus, sqlSorter);
 		
-		tvShowStatus.setLabelProvider(new SQLResultLabelProvider(reqQuery.getMode(), showStatus));
+		tvShowStatus.setLabelProvider(new QueryResultLabelProvider(reqQuery.getMode(), showStatus));
 		tvShowStatus.setContentProvider(new ArrayContentProvider());
 		
 		// 쿼리를 설정한 사용자가 설정 한 만큼 보여준다.
@@ -188,9 +188,9 @@ public class MySQLExtensionViewDialog extends Dialog {
 		final TadpoleResultSet trs = showExecutePlan.getDataList();
 		final SQLResultSorter sqlSorter = new SQLResultSorter(-999);
 		
-		SQLResultLabelProvider.createTableColumn(reqQuery, tvExecutePlan, showExecutePlan, sqlSorter, false);
+		TableUtil.createTableColumn(tvExecutePlan, showExecutePlan, sqlSorter);
 		
-		tvExecutePlan.setLabelProvider(new SQLResultLabelProvider(reqQuery.getMode(), showExecutePlan));
+		tvExecutePlan.setLabelProvider(new QueryResultLabelProvider(reqQuery.getMode(), showExecutePlan));
 		tvExecutePlan.setContentProvider(new ArrayContentProvider());
 		
 		// 쿼리를 설정한 사용자가 설정 한 만큼 보여준다.
